Ensure data type compatibility
The `GetExtendedAttribute` function correctly pulls a `NVARCHAR(MAX)` from `ExtendedAttributes`—but then incorrectly returns it as a `VARCHAR(MAX)`. That won't be an issue in many cases, but will be an issue if any UNICODE characters are used. This is easily fixed by ensuring both use `NVARCHAR(MAX)`.

This also quiets `SR0014`.