Moved new `UniqueKeyIndex` to `Utilities` schema
The new `UniqueKeyIndex` view (b09ca5d) is useful for debugging and reporting purposes. But it's _really_ slow. It should never be used in a production environment. To help emphasize this point, I've moved it to the `Utilities` schema.

As this was never deployed as part of a release, this isn't a breaking change.