Moved legacy `LeftRightRange` to `Utilities` schema
While the `LeftRightRange` view is a preexisting view, and thus this is a breaking change, it's not a view that is publicly documented, and it's exclusively used by the `CompressHierarchy` stored procedure—which is _also_ a preexisting object which was moved to the `Utilities` schema as part of a minor release under the same justification.

This helps keep the `dbo` schema focused on items which are publicly documented and expected to be used directly as part of the library.