Corrected `CompressHierarchy` name
The `CompressHierarchy` stored procedure was named `CompressHierarchy` in the file name and header, but `CompressTopics` in the actual code. That's not an intuitive name. I've renamed it for consistency and discoverability.

While this is a preexisting object, and is publicly documented, the public documentation has always referred to `CompressHierarchy`. As such, while this is technically a breaking change, it brings it inline with the publicly documented interface. Further, as it's a utility function that's not called from the product, I'm considering this acceptable for a minor release (as opposed to holding off on this fix until the next major release of OnTopic).