Moved legacy `Hierarchy` to `Utilities` schema, renamed `AdjacencyList`
While `Hierarchy` is a preexisting table, and thus this is a breaking change, it's not a table that is publicly documented, and it's exclusively used by the `GenerateNestedSet` stored procedure—which is _also_ a preexisting object which was moved to the `Utilities` schema as part of a minor release under the same justification.

As with other similar updates (e.g., e565f0e), this helps keep the `dbo` schema focused on items which are publicly documented and expected to be used directly as part of the library.