Added `FindAll()` overload to return all topics
There are a number of cases where we want all descendants of the current topic. The new `FindAll()` overload defines a default delegate that returns all topics.

Be aware that this still suffers from a bug due to the return type, `ReadOnlyTopicCollection`, limiting results to unique keys. That will be fixed in the next major version of OnTopic.