Switch to working off of a temporary table
Instead of working directly off of `Topics`, work instead off a light-weight temporary version of `#Topics` instead. Because this has no constraints or indexes, it is faster to work off of. In the future, this will also allow us to remove the `Stack_Top` column from the `Topics` table since it is exclusively needed for `GenerateNestedSet`.

In a subsequent commit, we'll need to reintroduce the insert back into `Topics`. This is currently broken in this version, as it doesn't account for the foreign key constraints, so we're not harming any functionality with this update.