Introduced new `SortOrder` column to `AdjacencyList`
An adjacency list, by default, has no sort order. But a nested set does. To help ensure the round trip from the nested set to an adjacency list and back to a nested set again maintains the integrity of topic ordering, I've introduced a new `SortOrder` column.

When populating the `AdjacencyList` from a nested set, the `RangeLeft` value can be used for the sort order. If populating the `AdjacencyList` from another source, this can just be set to e.g. `0` if there is no logical order available from the data set.