Honor new `SortOrder` when recreating the hierarchy
In the previous commit, a new `SortOrder` column was added to the `AdjacencyList` table to provide ordering hints not otherwise available in an adjacency list (61a1e95). In this commit, I wire that up to the actual generation of the nested set, so that it picks up the first child ordered by `SortOrder`, instead of the first `TopicID` ordered by `TopicID` (the default for the database).