Rearranged order of `#Topics` inserts
For historical reasons, `Stack_Top` is the first column of `Topics`. This isn't preferrable, but it's involved to rearrange columns after a table has been created. That isn't the case for the new `Topics` table, though, so we are able to place `Stack_Top` at the end of that declaration. This update reorders the `INSERT` statements to adhere to that order. This isn't strictly necessary, but it avoids confusion by maintaining a consistent order with the table columns.