Reenable constraints after insert
In a previous commit, I added code to insert the data from `#Topics` to `Topics`. As part of that, I needed to temporarily disable the foreign key constraints pointing to that table. These were supposed to be reenabled with that commit, but were inadvertantly overlooked from the cherry pick.

As part of that, I also removed a (now) duplicate disabling of `IDENTITY_INSERT`, since this has been moved to just before the `Topics` table is repopulated. This doesn't hurt anything to be duplicated, but it isn't necessary and could lead to confusion down the line.