Repopulate `Topics` in order of `TopicID`, not `RangeLeft`
In the original code that repopulates `Topics`, topics were recreated in the order of `RangeLeft` (d6b61c6). This makes sense given that we're working with a nested set. It doesn't make sense, however, given that the `TopicID` is the primary key. Given that, it will be (slightly) more performant to insert these in the order of `TopicID` since that's the physical order these are stored on the disk.

(This is only slightly faster because the indexes on `Topic` require much of this data to be rewritten regardless.)