Introduce transaction for `GenerateNestedSet`
We don't expect `GenerateNestedSet` to be run frequently, as it's a very invasive operation, and is likely only needed when the database is in an unusable state due to corruption of the hierarchy.

Nevertheless, we don't want to make matters worst in case there's an error that occurs while it's running—e.g., by deleting the `Topics` data, and then failing to reinsert it due to an issue with the data. While an unlikely scenario, this is easily avoided by introducing a transaction.

Further, we use the `SERIALIZABLE` isolation level to ensure that the tables are appropriately distanced from any read or write operations during this operation. We don't want, for instance, topics being created, deleted, or moved while the topic hierarchy is being regenerated.