Prevent `NULL` on `RangeRight`
Previously, `Topics.RangeRight` had to be nullable in order to facilitate the (uncommon!) use of `GenerateNestedSet`. Outside of that, we _never_ expect `RangeRight` to be `NULL`.

With `GenerateNestedSet` updated to instead operate off of a temporary `#Topics` table (dc72be2), this necessity no longer exists. As such, not only can we remove the `NOT NULL` option, but we can also remove a lot of defensive code that was put in to account for potential `NULL` values—even though, in practice, these should never have occurred.

This also provides better data integrity checks as there is now no way for the `RangeRight` to be set to `NULL`—a scenario we would never want to occur. If any code attempts to do that, an error will be thrown, and the record won't be inserted, even temporarily. That's a good thing!