Removed unnecessary `DateAdded` field from `AdjacencyList`
The legacy `DateAdded` field was not being actively used and served no apparent purpose. Given this, it is being removed. As part of this, I'm explicitly setting the columns in the `GenerateNestedSet` stored procedure to avoid any future confusion. (This is a best practice we should follow for all of our stored procedures.)