Refactored `GetAttributes` as user-defined, table-valued function
This doesn't impact backward compatibility, in that the function can still be called as a stored procedure. But it provides more flexibility in how the code is integrated with queries. For instance, it can now be used in a `CROSS APPLY` to retrieve attributes from `TopicID`s in another query.