Introduced new `FindTopicIDs` function
The `FindTopicIDs` function accepts a root `@TopicID`, `@AttributeKey`, and `@AttributeValue`, and finds all topics within the `@TopicID` that have an `AttributeKey` and `AttributeValue` matching the corresponding parameters.

Optionally, the search can be limited to either `Attributes` or `ExtendedAttributes` using the `@IsExtendedAttribute` parameter, or extended to a partial match with `@UsePartialMatch`.

Note: Functions don't actually support optional parameters. As such, `@IsExtendedAttribute` and `@UsePartialMatch` must be passed as values or as `DEFAULT`.