Updated `refactorlog` to reflect object changes
We've made a handful of refactorings such as renaming stored procedures, or moving objects to different schemas. By default, when using Visual Studio's _Schema Compare…_ feature, these objects will be dropped and recreated. That's not a huge issue outside of tables, but it's easier to follow the updates when they're tracked as changes. To support that, I've updated the `refactorlog` to track these. _Schema Compare…_ uses this to map old objects to new objects.

For some reason, Visual Studio's _Schema Compare…_ feature no longer seems to recognize or honor the `NewSchema` command, which appears to be a bug. That's unfortunate. Still, I'm keeping those in here if only for documentation.