Introduced `MappedTopicCache` and `MappedTopicCacheEntry`
Currently, a `ConcurrentDictionary<int, object>` is declared in the `TopicMappingService`. As we want to expand this to also track what `Relationships` flags were used to map the currently cached object, we're simplifying this by establishing a new `MappedTopicCache` object, which tracks a collection of `MappedTopicCacheEntry` objects. A `MappedTopicCacheEntry` provides access to both the `MappedTopic`—similar to the previous cache collection—as well as a `Relationships` property. This way, when an object from the cache is retrieved, it can be determined what relationships have already been mapped.