Introduced `MarkClean()` methods
Previously, to mark a collection as "clean", an implementor would need to loop through each and every `AttributeValue` and either a) directly set its `IsDirty` property to `false`, or b) call `AttributeValueCollection.SetValue()` with an explicit `isDirty` parameter value.

Now, implementors can simply call `MarkClean()`. This optionally accepts a `version` parameter to reset the `LastModified`; if not provided, it will default to `UtcNow`.

This can also be optionally applied to an individual `AttributeValue` by supplying the `name` parameter—though, in most circumstances, we'd expect it to be applied to the entire collection, which also handles tracking of deleted attributes.

This includes unit tests to validate the functionality.