Prefer `MarkClean()` over `IsDirty` on `Save()`
When calling `SqlTopicRepository.Save()`, call the new `AttributeValueCollection.MarkClean()` method (2ac653b) once the topic has been saved, instead of setting `AttributeValue.IsDirty` to `false` when evaluating the attributes.

There are three benefits to this. First, it ensures that the `Version` (`LastModified`) is reset. Second, it correctly resets deleted attribute tracking, which is independent of any on `AttributeValue`. Finally, it only resets the values after a successful call to the database. That means that if a `Save()` fails, dirty attributes will continue to be marked as `IsDirty`.