Check for duplicates based on key, not object
As `TypeCollection` is a `KeyedCollection`, checking to see if it contains the same `Type` is not useful—we need to confirm that it explicitly doesn't contain the same `Name`. Otherwise, two types with the same `Name` will result in a conflict. We haven't received any reports of this issue, nor encountered it ourselves, but it remains a possibility—and especially when e.g. overwriting models from `OnTopic.ViewModels`, where the names are enforced by convention.