Marked `FromList()` as deprecated
The `FromList()` method doesn't make much sense. It acts like a constructor method—but it's not static, so it effectively creates a duplicate object. Further, the same functionality is already accounted for by the constructor that accepts an `IList<T>` as a parameter, but which populates the current collection.

It's not entirely clear what the original thinking on this was, but it isn't actually used anywhere within the product or current implementations, and will be removed in the next major release of OnTopic.