Applied C# 9.0 pattern matching for equality checks against constants
Replaced all instances of `==`, `!=`, and `.Equals()` with `is` or `is not` where the value being compared against is a constant. This builds upon the previous commit (6ac12e8), which applied pattern matching for cases where we were previously using `&&` or `||`.

The one exception to this change are cases where a) a case insensitive comparison was being performed, and b) that was necessary. For example, currently `ContentType` is case insensitive. And while it's really unexpected for a `ContentType` to _not_ use the standard casing, it's something we've permitted to date. As such, comparisons against a `ContentType` (e.g., topic.ContentType.Equals("List", StringComparison.OrdinalIgnoreCase)`) will continue to be used, even though "List" is a constant, since `is` is (obviously) not case sensitive.