Updated `TopicId` lookup to be case insensitive
The application of C# 9.0 pattern matching (5b1c685) exposed the pattern of using case-insensitive topic and attribute keys, including for the content type. It's unclear if this is really necessary—it shouldn't be—but so long as we're maintaining that, we should be consistent. This picks up at least one case where we weren't using the case-insensitive comparison and it might matter—though, it's worth noting, that this rule has been inconsistently applied, and there remain other exceptions.