Implemented implicit types for `new` expressions
With C# 9.0, the `new` keyword only needs a type declaration if the type is otherwise unknown (e.g., `var`) or ambiguous (e.g., with an interface, or overloaded method). Otherwise, it can just use `new()`. This is especially useful for property initializers, which can be incredibly repetitive.

Note: This apparently cannot be done for generics (e.g., `new T()` cannot become `new()`). I'll need to look into that, but for now it's useful to know that it doesn't work.