Fixed logic for `validationDelegate`
In a previous commit, I fixed a bug on when the `validationDelegate()` was called (64813c7). This resulted in more intuitive and flexible behavior that aligns with the documented functionality. In doing so, however, it potentially breaks implementations—including that in the `MenuViewComponentBase<T>`.

As the `validationDelegate()` is now being called correctly on each individual topic, instead of on the parent topic, we need to evaluate the content type of the parent topic in order to filter out children of the `PageGroup` content type. While I was at it, I updated the logic to use C# 9.0's pattern matching syntax.