Extended 'validationDelegate` to remove `List` topics
There is no scenario where we expect a `List` topic to be included in the navigation—or, generally, for the children of `List` topics to be included, and especially since a `List` cannot be navigated to. If we did, that would be a highly specialized scenario, and would necessitate a custom `validationDelegate()` to handle that logic. As such, we should strip `List` topics out of the default `MenuViewComponentBase<T>` so that they don't need to be explicitly filtered out at the view level, as they are today.

(This also makes a minimal gain in terms of memory and performance since these unnecessary topics don't need to be mapped or cached in memory. That's likely negligible, of course.)