Introduced a new `IsIndexed` property
The `IsIndexed` property allows a view to implement e.g. a table of contents for either the categories or the items on the page, while giving the editor control over whether or not that option is enabled. Not all implementors will expose this option in the editor, and it's up to the view designer to determine how it will be implemented in the markup. Nevertheless, it's a common enough feature that having it built-in to the view model offers a lot of flexibility.