Introduced a new `IndexLabel` property
In the previous commit, I introduced a new `IsIndexed` property, which allows an optional table of contents to be exposed on a `ContentList` page (ebbeeff). This commit introduces a new `IndexLabel` which, optionally, allows the label used for any such table of contents to be set. By default, it will be "Contents". As with `IsIndexed`, it's up to implementors as to whether this is exposed in the editor, and how it will be honored in the view.