Update to Code Analysis 3.3.1
Despite only being a minor patch, this is a critical update in that it resolves a number of false positives introduced when using C# 9.0. That said, it doesn't resolve _all_ of them, so hopefully we'll see another update soon. (Notably, some instances of `is not null` are not being read as a null check, and valid cases of potential null references are still being flagged.)