Allow class names to end in `Attribute`
Normally, class names should not end in `Attribute` unless they derive from `Attribute`. That's a good rule. Unfortunately, OnTopic has preexisting naming conventions with `Attribute` representing an attribute descriptor, and class names named after content type descriptors. As a result, there are classes that need to be named e.g. `BooleanAttribute` or `TextAttribute`. Perhaps these should be named `*AttributeDescriptor`, but that's a breaking change and will need to be reevaluated in a future major release. For now, supressing CA1711 resolves this warning.