Remove unnecessary null checks
There are a number of areas in the code where we perform a null check against references that should never be null due e.g. to preexisting guard clauses. This removes those unnecessary checks, and resolves most CA1508 warnings.

Note: There are a few instances which remain, due to conflicts with CS8602, and will need to be addressed later.