Renamed attributes to align with corresponding properties
Typically, constructor parameters which set properties should have identifiers that match those properties. This is a best practice that Ignia normally follows. We don't, however, for some cases with attributes, in order to avoid seemingly redundant property names. For example, `FilterByAttribute.Key` makes more sense than `FilterByAttribute.AttributeKey` since attribute is implied by the name of the class—but to keep IntelliSense explicit for parameters, the full `attributeKey` was used.

There's an argument that perhaps the attributes and the properties should use the full name, but that's a potentially breaking change. To limit the impact, the parameter name is being updated. (Technically, renaming the parameter is also a potentially breaking change, but as we have insight into current implementers, we know that no one is currently explicitly referencing the parameter by name.)

This resolves CA1019.