Throw `InvalidOperationException` instead of `Exception`
There are a few scenarios where we were throwing `Exception` instead of more specific exceptions. These were not documented or expected calls, and we have a fair amount of confidence that implementers are not catching `Exception` in order to handle these specific errors. In fact, generally, these exceptions are thrown to warn developers of unexpected scenarios usually related to the data model, such as infinite loops—i.e., we don't expect these to occur in production environments.

Given this, it doesn't really matter _which_ exception they throw, and while swapping out an exception is potentially a breaking change, we have confidence in these scenarios that they won't actually break anything, as these scenarios represent a design flaw and would be caught during testing. (Indeed, technically, a couple of these could probably be replaced with code analyzers in the future for design time and compile time support.)

This resolves most instances of CA2201. Note that there remain a couple of false positives of that warning that can't be removed due to generic constraints requiring the _possibility_ of an `Exception` (in order to restrict the generic to any exception type) even though, in practice, we would never expect a caller to pass in `Exception` as the type parameter. We may revisit that in the future by supressing those particular warnings.