Suppress warning for unexpected throwing of `Exception`
CA2201 occurs when an exception of type `Exception` is thrown, as more specific exceptions should be preferred. In practice, we don't expect this to occur on the `Contract` class, either, but due to limitations of generic type constraints, there isn't a way to restrict the type parameter to a derivative of `Exception` while not permitting `Exception` to potentially be passed. Ideally, this warning would occur at the source, when the generic type is defined, but that's not supported by the code analyzer—and would likely be quite complicated (if not impossible) to create. As such, instead, we're suppressing this warning.

Should we decide this is being abused, we might consider writing a custom code analyzer for this scenario, or adding a runtime exception if the `typeof(T)` is `typeof(Exception)`. As this is an internal library, however, we don't anticipate that being an issue.