Convert `List<>` to `Collection<>`
CA1002 advises using the more comprehensive and familiar `Collection<>` type for public members over the faster, but more bare bones, `List<>` type. This is a potentially breaking change so we'll need to hold off until the next version to make that change to any libraries published as NuGet packages. For internal tests, however, we can make that change now in order to validate future compatibility.

As expected, this doesn't impact any of our unit tests since the expected members supported by `List<>` are also supported by `Collection<>`—and, in fact, as many of these relate to the `ITopicMappingService`, it's exclusively interacting with them as `IList`, which both support.

In practice, we could probably convert the remaining items over to `Collection<>` due to this, without worrying about introducing a potentially breaking change. But as this could cause issues with e.g. casting or variable compatibility, we'll wait until OnTopic 5.0. For those outliers, I introduced a global suppression (for `OnTopic`) and a single local suppression (for `OnTopic.Tests`).