Remove C# 9.0 "or" pattern combinator
The `NetAnalyzers` code analysis library doesn't yet understand C# 9.0's new "or" pattern combinator and, thus, throws a CA1062 false positive when a parameter is validated via a guard clause with the form, `parameter is null or …` (see dotnet/roslyn-analyzers#4496). Until that issue is resolved, I'm falling back to the legacy `paramer is null || parameter is …` format, which `NetAnalyzers` understands.