Suppress check for injection vulnerabilities
Due to a bug in the `Microsoft.CodeAnalysis.NetAnalyzers`, a number of CA30XX analyzers throw exceptions, cluttering up the warnings in Visual Studio. A fix has been proposed for these (dotnet/roslyn-analyzers#4495), and will hopefully be available in the next release of the .NET SDK. Until then, these errors are being suppressed.

Unfortunately, it doesn't appear to be possible to suppress these warnings at the source level or even via a global suppressions file. Instead, they need to be suppressed at the project level. That's unfortunate, as it means they're not targeted toward the one scenario which trips these exceptions (i.e., the public setter for `TopicController.CurrentTopic`).

We'll want to remove these as soon as this issue is resolved.