Supress suggestion to use C# 9.0's "or" pattern combinator
In a previous commit, we removed use of C# 9.0's "or" pattern combinator due to a bug in `NetAnalyzers` which yielded a false positive for CA1062 (6affe62). This results in an IDE0078, suggesting the use of the "or" pattern combinator. To temporarily mitigate that, I'm suppressing that warning. This should be revisited later.