Ensure `ModelType` has default value
In practice, we expect the `ModelType` enum to always have a value, based on how it is designed. However, enums default to 0 when initialized, and so not havine a `None` assigned to 0 can open up for unexpected bugs, and especially in downlevel code. To avoid any potential confusion, a `None` option is added. This mitigates CA1008.