Establish `AssemblyInfo` class for `TestDoubles` project
Normally, all of our projects have an `AssemblyInfo` class for storing properties setup via the default template. As the `TestDoubles` project was created later, with a newer template, it didn't include one. Most properties are now handled via the `csproj` file so that's not a huge issue. One property that remains needed, however, is the `CLSCompliant` attribute, which confirms that the assembly conforms to CLS conventions. This resolves CA1014. While I was at it, I also assigned the correct GUID for the assembly, as previously assigned in the `sln` file.