Suppress CA1024 as appropriate
CA1024 recommends avoiding `Get()` methods without parameters in preference for properties. That's good advice. But there are a few cases where it doesn't make sense. In these cases, a value is being dynamically calculated based on context. This isn't terribly time consuming, but it is certainly more than just returning a field. And the value could change based on contextual updates outside the scope of the current object. Further, there are conventions built into e.g. the `TopicMappingService` which recognize `Get{Attribute}()` methods as alternatives to `{Attribute} { get; }` accessors when mapping DTOs. In these cases, these methods are deliberate and should be persisted.