Installed NuGet version of `NetAnalyzers`
With the migration from `FxCopAnalyzers` to `NetAnalyzers`, we no longer _need_ to use the NuGet package, as the analyzers are now built into the SDK (7733c0c). Unfortunately, that also means that they're going to lag behind the latest fixes.

When we originally tried installing the NuGet package (7733c0c), we got duplicate warnings, and so had relied on the SDK implementation. That appears to have been subsequently resolved. As such, I'm moving back to the NuGet implementation to ensure we're always using the latest version.