Introduced a new `BusinessLogicCache`
If `Topic` or a derivative contains a property with the same name as an attribute, it should be called whenever an attribute is updated in order to ensure any business logic is enforced. To do this, we need to keep track of the initial `AttributeValue` (before executing the business logic) as well as whether the business logic has been triggered already (to avoid an infinite loop).

We'd like to move back toward a read-only `AttributeValue` class, which was originally intended as a value object. To do that, we need to start by removing state tracking from it. That includes the `EnabledBusinessLogic` property.

To replace this, we need to establish a different way of tracking whether or not the business logic is currently in progress, and to maintain access to the original `AttributeValue` in order to maintain optional attributes that `InsertItem()` and `SetItem()` aren't otherwise aware of (such as `IsDirty` and `LastModified`). To do this, we implement a new `Dictionary<string, AttributeValue?>` as a local cache.

If a record is in this cache, that means it has already gone through its initial evaluation, and is now having its business logic enforced. This cache can be checked to get the `AttributeValue` from that initial evaluation, and to confirm the current state of enforcement.

As this is a bit unintuitive, I've included plenty of documentation, even though it's a `private` member.

In subsequent commits, we'll be wiring up this logic elsewhere in the `AttributeValueCollection` class.