Add logic for maintaining `BusinessLogicCache`
Previously, the `enableBusinessLogic` parameter of `SetValue()` was evaluated at the top so that it could be added to the `AttributeValue` objects when they were created. With the introduction of the `BusinessLogicCache`, that no longer works as the `BusinessLogicCache` needs to maintain a copy of the `AttributeValue` and, thus, that logic can't happen until _after_ the `AttributeValue` has been created.

In addition to this, the cache must be populated _prior_ to calling `Add()` or `InsertAt()`; otherwise, an infinite loop will occur since the underlying `SetItem()` and `InsertItem()` methods will enforce the business logic; if the business logic cache isn't established yet, they won't be able to detect that the business logic has already been triggered. Remember, the whole purpose of the `BusinessLogicCache` is to keep track of this state so that `AttributeValue` doesn't, while still ensuring that `SetItem()` and `InsertItem()` can determine the current state.

Given this, we not only need to move the `enforceBusinessLogic` evaluate to the bottom, but we need to then move call to add (`Add(…)`) or replace (`this[…]`) the `AttributeValue` below that. This requires some minor rearranging of the variables so that we can establish the `updatedAttributeValue` at the top, set it within each condition, and finally add it after the `BusinessLogicCache` has been established.