Handle business logic within `SetValue()`
Previously, the `EnforceBusinessLogic()` method not only triggered the business logic, but also handled returning the original `AttributeValue` back to `SetItem()` or `InsertItem()`. This was necessary to ensure that any optional parameters sent to `SetValue()`—such as `isDirty` or `version`—were maintained, since `SetItem()` and `InsertItem()` wouldn't otherwise be aware of them.

This is no longer necessary since the `BusinessLogicCache` now offers access to the original `AttributeValue` from before `EnforceBusinessLogic()` called out to the underlying property. As such, for call to `SetValue()` triggered by the `Topic` properties, we can now detect that it's on the return trip from calling the business logic, and resurrect that original `AttributeValue`, with any optional parameters already set to its properties, and persist that directly.

This is handy as it not only simplifies the logic, but also prevents us from needing to create a new `AttributeValue` for the business logic, only to discard it at the last minute. Now, we can just reinsert the original `AttributeValue`.

(We are assuming now, just as we assumed previously, that the `AttributeValue.Value` will be the same on the new version and the original version, since we don't expect calling a property setter to _alter_ the value. If the value is invalid, and _exception_ will be thrown, at which point this code won't execute. As such, the only differences between the initial `AttributeValue` and the `AttributeValue` that would have been created by a second call to `SetValue()` should be those state properties that we weren't able to relay through the `EnforceBusinessLogic()` behavior, such as `IsDirty` and `LastModified`.)

This allows us to get rid of the `out` parameter on `EnforceBusinessLogic()` since `SetItem()` and `InsertItem()` are now operating off of the original version, and don't need the new version to be swapped out with the original.