Remove legacy `EnforceBusinessLogic` property
With the business logic state now successfully migrated over to the new `BusinessLogicCache` within `AttributeValueCollection`, there's no longer a need to maintain the `EnforceBusinessLogic` property on `AttributeValue`. This brings us one step closer to being able to return `AttributeValue` to a read-only value object—and, in the meanwhile, has effectively cleaned up the logic within `AttributeValueCollection`.