Remove `DeletedAttributes` when inserting new attributes
If an `AttributeValue` is removed, its key will be added to the new `DeletedAttributes` collection for tracking (54eab63). If an `AttributeValue` is subsequently added with the same `Key`, we no longer want to track that as a deletion. This is effectively what should be viewed as a replacement, even though it took place in two steps.

To do that, we'll check the `DeletedAttributes` collection on `InsertItem()` and `SetItem()`.

While I was at it, I also introduced missing `Contract.Requires()` for the `item`, since those methods will otherwise throw `NullReferenceException`s when attempting to dereference the `item` parameter (e.g., for `item.Key` references).